# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type GeminiChatMessageRole = typeof GeminiChatMessageRole[keyof typeof GeminiChatMessageRole];


export const GeminiChatMessageRole = {
  user: 'user',
  model: 'model',
} as const;

export interface GeminiChatMessage {
  role: GeminiChatMessageRole;
  text: string;
}

export type GeminiChatRequestMode = typeof GeminiChatRequestMode[keyof typeof GeminiChatRequestMode];


export const GeminiChatRequestMode = {
  irfan: 'irfan',
  elena: 'elena',
  together: 'together',
} as const;

export type GeminiChatRequestAgentsItem = typeof GeminiChatRequestAgentsItem[keyof typeof GeminiChatRequestAgentsItem];


export const GeminiChatRequestAgentsItem = {
  irfan: 'irfan',
  elena: 'elena',
} as const;

export interface GeminiChatRequest {
  message: string;
  mode?: GeminiChatRequestMode;
  agents?: GeminiChatRequestAgentsItem[];
  history?: GeminiChatMessage[];
  systemInstruction?: string;
}

export type GeminiChatResponseAgent = typeof GeminiChatResponseAgent[keyof typeof GeminiChatResponseAgent];


export const GeminiChatResponseAgent = {
  irfan: 'irfan',
  elena: 'elena',
} as const;

export type GeminiChatResponseContributorsItem = typeof GeminiChatResponseContributorsItem[keyof typeof GeminiChatResponseContributorsItem];


export const GeminiChatResponseContributorsItem = {
  irfan: 'irfan',
  elena: 'elena',
} as const;

export interface GeminiChatResponse {
  text: string;
  agent: GeminiChatResponseAgent;
  collaboration: boolean;
  contributors?: GeminiChatResponseContributorsItem[];
  degraded?: string;
}