# lib/api-zod/src/generated/types/geminiChatMessage.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GeminiChatMessageRole } from './geminiChatMessageRole';

export interface GeminiChatMessage {
  role: GeminiChatMessageRole;
  text: string;
}