# lib/api-zod/src/generated/types/geminiChatRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GeminiChatMessage } from './geminiChatMessage';
import type { GeminiChatRequestAgentsItem } from './geminiChatRequestAgentsItem';
import type { GeminiChatRequestMode } from './geminiChatRequestMode';

export interface GeminiChatRequest {
  message: string;
  mode?: GeminiChatRequestMode;
  agents?: GeminiChatRequestAgentsItem[];
  history?: GeminiChatMessage[];
  systemInstruction?: string;
}