# lib/api-zod/src/generated/types/geminiChatResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GeminiChatResponseAgent } from './geminiChatResponseAgent';
import type { GeminiChatResponseContributorsItem } from './geminiChatResponseContributorsItem';

export interface GeminiChatResponse {
  text: string;
  agent: GeminiChatResponseAgent;
  collaboration: boolean;
  contributors?: GeminiChatResponseContributorsItem[];
  degraded?: string;
}